Personalización de mensajes de error/success

+++ Proyecto FINAL/proyecto/src/main/java/com/domination/proyecto/controllers/SalaController.java
@@ -7,6 +7,7 @@ import com.domination.proyecto.models.Reserva;
 import com.domination.proyecto.services.SalaService;
 import com.domination.proyecto.services.SucursalService;
 import com.domination.proyecto.services.ReservaService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -72,29 +73,60 @@ public class SalaController {
     }
 
     @PostMapping("/create")
-    public String createSala(@ModelAttribute Sala sala, Model model) {
-        salaService.save(sala);
-        model.addAttribute("message", "La Sala ha sido creada exitosamente");
-        return "redirect:/salas/salasDisponibles/" + sala.getSucursal().getIdSucursal();
+    public String createSala(@ModelAttribute Sala sala, Model model, HttpSession session) {
+        try {
+            salaService.save(sala);
+            session.setAttribute("Exito", true);
+            session.setAttribute("mensaje", "La Sala ha sido creada exitosamente");
+            return "redirect:/salas/salasDisponibles/" + sala.getSucursal().getIdSucursal();
+        } catch (Exception e) {
+            session.setAttribute("Exito", true);
+            session.setAttribute("mensaje", "Error al crear la Sala");
+            return "redirect:/salas/salasDisponibles/" + sala.getSucursal().getIdSucursal();
+        }
     }
 
     @PostMapping("/update")
-    public String updateSala(@RequestParam("idSucursal") int idSucursal,@ModelAttribute Sala sala, Model model) {
-        Sucursal sucursal = sucursalService.findByIdSucursal(idSucursal)
-                                           .orElseThrow(() -> new ObjectNotFoundException("Sucursal no encontrada con id: " + idSucursal));
-        sala.setSucursal(sucursal);
-        salaService.save(sala);
-        model.addAttribute("message", "La sala ha sido actualizada exitosamente");
-        return "redirect:/salas/salasDisponibles/"+ sala.getSucursal().getIdSucursal();
+    public String updateSala(@RequestParam("idSucursal") int idSucursal,@ModelAttribute Sala sala, Model model, HttpSession session) {
+        try {
+            Sucursal sucursal = sucursalService.findByIdSucursal(idSucursal)
+                    .orElseThrow(() -> new ObjectNotFoundException("Sucursal no encontrada con id: " + idSucursal));
+            sala.setSucursal(sucursal);
+            salaService.save(sala);
+            session.setAttribute("Exito", true);
+            session.setAttribute("mensaje", "La Sala ha sido actualizada exitosamente");
+            return "redirect:/salas/salasDisponibles/" + sala.getSucursal().getIdSucursal();
+        } catch (Exception e) {
+            session.setAttribute("Exito", true);
+            session.setAttribute("mensaje", "Error al actualizar la Sala");
+            return "redirect:/salas/salasDisponibles/" + sala.getSucursal().getIdSucursal();
+        }
     }
 
     @PostMapping("/delete")
-    public String deleteSala(@RequestParam("idSala") int idSala, Model model) {
-        Sala sala = salaService.findById(idSala)
-                               .orElseThrow(() -> new ObjectNotFoundException("Sala no encontrada con id: " + idSala));
- //       reservaService.deleteReservasBySala(sala); // Si hay reservas relacionadas
-        salaService.deleteById(idSala);
-        model.addAttribute("message", "La sala ha sido eliminada exitosamente");
-        return "redirect:/salas/salasDisponibles/" + sala.getSucursal().getIdSucursal();
+    public String deleteSala(@RequestParam("idSala") int idSala, Model model, HttpSession session) {
+        try {
+            Sala sala = salaService.findById(idSala)
+                                   .orElseThrow(() -> new ObjectNotFoundException("Sala no encontrada con id: " + idSala));
+            /* Versión del algoritmo que verifica si la sala tiene reservas asociadas antes de eliminarla
+
+            List<Reserva> reservas = reservaService.findReservasBySala(sala);
+            if (reservas.isEmpty()) {
+                salaService.deleteById(idSala);
+                model.addAttribute("mensaje", "La Sala ha sido eliminada exitosamente");
+                return "redirect:/salas/salasDisponibles/" + sala.getSucursal().getIdSucursal();
+            } else {
+                model.addAttribute("mensaje", "No se puede eliminar la Sala porque tiene reservas asociadas");
+                return "redirect:/salas/salasDisponibles/" + sala.getSucursal().getIdSucursal();
+            }*/
+            salaService.deleteById(idSala);
+            session.setAttribute("Exito", true);
+            session.setAttribute("mensaje", "La sala ha sido eliminada exitosamente");
+            return "redirect:/salas/salasDisponibles/" + sala.getSucursal().getIdSucursal();
+        } catch (Exception e) {
+            session.setAttribute("Exito", true);
+            session.setAttribute("mensaje", "Error al eliminar la Sala");
+            return "redirect:/salas/salasDisponibles/" + idSala;
+        }
     }
 }
